==> 3_Tuning Evaluation.ipynb <==
# 3. Tuning Evaluation (for Tuning part 1 output)

import pandas as pd

Please adjust this cell to change files to be viewed

MODEL_NAME = 'RF'
DTYPE = 'TFIDF_Preprocessed'

file = pd.read_csv(f'./Tuning/Tuning - {MODEL_NAME} {DTYPE}.csv')

# View tuning results
file.sort_values(by = 'accuracy', ascending = False).head(60)
